requirements.txt
fastapi
uvicorn
sqlmodel
pymysql
python-dotenv
requests
openai
pymupdf